+++ packages/astro/src/content/vite-plugin-content-virtual-mod.ts
@@ -94,7 +94,7 @@ export function astroContentVirtualModPlugin({
 /**
  * Generate a map from a collection + slug to the local file path.
  * This is used internally to resolve entry imports when using `getEntry()`.
- * @see `src/content/virtual-mod.mjs`
+ * @see `content-module.template.mjs`
  */
 export async function getStringifiedLookupMap({
 	contentPaths,
@@ -166,6 +166,16 @@ export async function getStringifiedLookupMap({
 						fileUrl: pathToFileURL(filePath),
 						contentEntryType,
 					});
+					if (lookupMap[collection]?.entries?.[slug]) {
+						throw new AstroError({
+							...AstroErrorData.DuplicateContentEntrySlugError,
+							message: AstroErrorData.DuplicateContentEntrySlugError.message(collection, slug),
+							hint:
+								slug !== generatedSlug
+									? `Check the \`slug\` frontmatter property in **${id}**.`
+									: undefined,
+						});
+					}
 					lookupMap[collection] = {
 						type: 'content',
 						entries: {