+++ packages/astro/src/core/config/schema.ts
@@ -24,6 +24,7 @@ const ASTRO_CONFIG_DEFAULTS: AstroUserConfig & any = {
 		serverEntry: 'entry.mjs',
 		redirects: true,
 		inlineStylesheets: 'never',
+		split: false,
 	},
 	compressHTML: false,
 	server: {
@@ -120,6 +121,8 @@ export const AstroConfigSchema = z.object({
 				.enum(['always', 'auto', 'never'])
 				.optional()
 				.default(ASTRO_CONFIG_DEFAULTS.build.inlineStylesheets),
+
+			split: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.build.split),
 		})
 		.optional()
 		.default({}),
@@ -279,6 +282,8 @@ export function createRelativeSchema(cmd: string, fileProtocolRoot: URL) {
 					.enum(['always', 'auto', 'never'])
 					.optional()
 					.default(ASTRO_CONFIG_DEFAULTS.build.inlineStylesheets),
+
+				split: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.build.split),
 			})
 			.optional()
 			.default({}),