+++ packages/astro/test/events.test.js
@@ -385,6 +385,14 @@ describe('Events', () => {
 			expect(payload.config.integrations.length).to.equal(0);
 		});
 
+		it('finds names for integration arrays', () => {
+			const config = {
+				integrations: [{ name: 'foo' }, [{ name: 'bar' }, { name: 'baz' }]],
+			};
+			const [{ payload }] = events.eventCliSession({ cliCommand: 'dev' }, config);
+			expect(payload.config.integrations).to.deep.equal(['foo', 'bar', 'baz']);
+		});
+
 		it('includes cli flags in payload', () => {
 			const config = {};
 			const flags = {